feat: add validation to forms

+++ app/lib/actions.ts
@@ -4,6 +4,7 @@ import { sql } from "@vercel/postgres";
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
 
+// 定义好数据的模型，以及校验的信息
 const FormSchema = z.object({
   id: z.string(),
   customerId: z.string({
@@ -18,6 +19,7 @@ const FormSchema = z.object({
   date: z.string(),
 });
 
+// 状态类型
 export type State = {
   errors?: {
     customerId?: string[];
@@ -27,11 +29,9 @@ export type State = {
   message?: string | null;
 };
 
+// 新增
 const CreateInvoice = FormSchema.omit({ id: true, date: true });
-
 export async function createInvoice(prevState: State, formData: FormData) {
-  console.log("createInvoice ==>", formData);
-
   const validatedFields = CreateInvoice.safeParse({
     customerId: formData.get("customerId"),
     amount: formData.get("amount"),
@@ -66,27 +66,60 @@ export async function createInvoice(prevState: State, formData: FormData) {
   redirect("/dashboard/invoices");
 }
 
+// 更新
 const UpdateInvoice = FormSchema.omit({ id: true, date: true });
 export async function updateInvoice(id: string, formData: FormData) {
-  const { customerId, amount, status } = UpdateInvoice.parse({
+  const validatedFields = UpdateInvoice.safeParse({
     customerId: formData.get("customerId"),
     amount: formData.get("amount"),
     status: formData.get("status"),
   });
 
-  const amountInCents = amount * 100;
+  if (!validatedFields.success) {
+    return {
+      errors: validatedFields.error.flatten().fieldErrors,
+      message: "Missing Fields. Failed to Update Invoice.",
+    };
+  }
 
-  await sql`
+  const { customerId, amount, status } = validatedFields.data;
+  const amountInCents = amount * 100;
+  try {
+    await sql`
       UPDATE invoices
       SET customer_id = ${customerId}, amount = ${amountInCents}, status = ${status}
       WHERE id = ${id}
     `;
+  } catch (error) {
+    return { message: "Database Error: Failed to Update Invoice." + error };
+  }
 
   revalidatePath("/dashboard/invoices");
   redirect("/dashboard/invoices");
 }
 
+// 删除
+const DeleteInvoice = FormSchema.omit({
+  customerId:true,
+  amount:true,
+  status:true,
+  date:true,
+})
 export async function deleteInvoice(id: string) {
-  await sql`DELETE FROM invoices WHERE id = ${id}`;
+  const validatedFields = DeleteInvoice.safeParse({
+    id
+  })
+
+  if (!validatedFields.success) {
+    return {
+      errors: validatedFields.error.flatten().fieldErrors,
+      message: "Missing Fields:id. Failed to Delete Invoice.",
+    };
+  }
+  try {
+    await sql`DELETE FROM invoices WHERE id = ${id}`;
+  } catch (error) {
+    return { message: "Database Error: Failed to Update Invoice." + error };
+  }
   revalidatePath("/dashboard/invoices");
 }
